Добавлена проверка, находится ли компонент в состоянии редактирования.
git-svn-id: https://energine.googlecode.com/svn/trunk@1951 acdd3125-bb50-0410-93b3-95ffde4b7c7d

+++ core/modules/share/components/LinkingEditor.class.php
@@ -27,10 +27,14 @@
      */
     public function __construct($name, $module,   array $params = null) {
         parent::__construct($name, $module,  $params);
-        $this->isEditable = $this->document->isEditable();
-        $this->setProperty('exttype', 'feededitor');
-        if(!in_array($this->getState(),  array('up', 'down')) && isset($_COOKIE[md5($this->getName())])){
-            E()->getResponse()->deleteCookie(md5($this->getName()));
+        if(!($this->isEditable = $this->document->isEditable()) && ($this->getState() == self::DEFAULT_STATE_NAME)){
+           $this->disable();
+        }
+        else {
+            $this->setProperty('exttype', 'feededitor');
+            if(!in_array($this->getState(),  array('up', 'down')) && isset($_COOKIE[md5($this->getName())])){
+                E()->getResponse()->deleteCookie(md5($this->getName()));
+            }
         }
     }
